Break up DesignerObject into separate files
