+++ designer/__init__.py
@@ -2,7 +2,7 @@ import pygame
 from designer.Director import *
 from designer.helpers import *
 from designer.Animation import *
-from designer.DesignerObject import *
+from designer.objects import *
 from designer.colors import *
 
 GLOBAL_DIRECTOR = None
