+++ designer/objects/__init__.py
@@ -1,0 +1,14 @@
+"""
+The basic sprite ("DesignerObject") and subclass sprites (image, circle, rectangle, text, various widgest, etc.).
+"""
+
+from designer.objects.designer_object import DesignerObject
+from designer.objects.arc import Arc, arc
+from designer.objects.circle import Circle, circle
+from designer.objects.ellipse import Ellipse, ellipse
+from designer.objects.group import DesignerGroup, group
+from designer.objects.image import Image, image
+from designer.objects.line import Line, line
+from designer.objects.rectangle import Rectangle, rectangle
+from designer.objects.shape import Shape, shape
+from designer.objects.text import text
