+++ setup.py
@@ -14,7 +14,7 @@ setuptools.setup(
     version='0.0.88',
     python_requires='>=3.6',
     author='krishols, acbart',
-    packages=['designer'],
+    packages=['designer', 'designer.objects', 'designer.utilities'],
     package_data={'designer': ['data/colors.json']},
     author_email='kris@udel.edu',
     description='Student-friendly and evidence-based visual graphics library.',
